# [Enhancement] Optimize iceberg mor performance of iceberg equality delete