--- a/be/src/exec/hdfs_scanner.h
+++ b/be/src/exec/hdfs_scanner.h
@@ -17,7 +17,6 @@
 #include <atomic>
 #include <boost/algorithm/string.hpp>
 
-#include "exec/mor_processor.h"
 #include "exec/pipeline/scan/morsel.h"
 #include "exprs/expr.h"
 #include "exprs/expr_context.h"
@@ -189,6 +188,11 @@ struct HdfsScannerParams {
     // partition conjunct, used to generate partition columns
     std::vector<ExprContext*> partition_values;
 
+    std::vector<SlotDescriptor*> extended_col_slots;
+    std::vector<int> extended_col_index_in_chunk;
+    std::vector<int> index_in_extended_columns;
+    std::vector<ExprContext*> extended_col_values;
+
     // min max conjunct for filter row group or page
     // should clone in scanner
     std::vector<ExprContext*> min_max_conjunct_ctxs;
@@ -205,8 +209,6 @@ struct HdfsScannerParams {
 
     const TIcebergSchema* iceberg_schema = nullptr;
 
-    const TIcebergSchema* iceberg_equal_delete_schema = nullptr;
-
     bool is_lazy_materialization_slot(SlotId slot_id) const;
 
     std::shared_ptr<TPaimonDeletionFile> paimon_deletion_file = nullptr;
@@ -218,7 +220,6 @@ struct HdfsScannerParams {
     bool can_use_any_column = false;
     bool can_use_min_max_count_opt = false;
     bool orc_use_column_names = false;
-    MORParams mor_params;
 
     int64_t connector_max_split_size = 0;
 };
@@ -252,6 +253,11 @@ struct HdfsScannerContext {
 
     // partition column value which read from hdfs file path
     std::vector<ColumnPtr> partition_values;
+
+    // extended column
+    std::vector<ColumnInfo> extended_columns;
+
+    std::vector<ColumnPtr> extended_values;
 
     // scan range
     const THdfsScanRange* scan_range = nullptr;
@@ -313,6 +319,8 @@ struct HdfsScannerContext {
     // otherwise update partition column in chunk
     void append_or_update_partition_column_to_chunk(ChunkPtr* chunk, size_t row_count);
     void append_or_update_count_column_to_chunk(ChunkPtr* chunk, size_t row_count);
+
+    void append_or_update_extended_column_to_chunk(ChunkPtr* chunk, size_t row_count);
 
     // if we can skip this file by evaluating conjuncts of non-existed columns with default value.
     StatusOr<bool> should_skip_by_evaluating_not_existed_slots();
@@ -383,7 +391,6 @@ private:
     std::atomic<bool> _closed = false;
     Status _build_scanner_context();
     void update_hdfs_counter(HdfsScanProfile* profile);
-    Status _init_mor_processor(RuntimeState* runtime_state, const MORParams& params);
 
 protected:
     HdfsScannerContext _scanner_ctx;
@@ -397,8 +404,6 @@ protected:
     std::shared_ptr<io::CacheInputStream> _cache_input_stream = nullptr;
     std::shared_ptr<io::SharedBufferedInputStream> _shared_buffered_input_stream = nullptr;
     int64_t _total_running_time = 0;
-
-    std::shared_ptr<DefaultMORProcessor> _mor_processor;
 };
 
 } // namespace starrocks