--- a/be/src/runtime/descriptors.h
+++ b/be/src/runtime/descriptors.h
@@ -241,7 +241,6 @@ public:
     ~IcebergTableDescriptor() override = default;
     bool has_partition() const override { return false; }
     const TIcebergSchema* get_iceberg_schema() const { return &_t_iceberg_schema; }
-    const TIcebergSchema* get_iceberg_equal_delete_schema() const { return &_t_iceberg_equal_delete_schema; }
     bool is_unpartitioned_table() { return _partition_column_names.empty(); }
     const std::vector<std::string>& partition_column_names() { return _partition_column_names; }
     const std::vector<std::string> full_column_names();
@@ -252,7 +251,6 @@ public:
 
 private:
     TIcebergSchema _t_iceberg_schema;
-    TIcebergSchema _t_iceberg_equal_delete_schema;
     std::vector<std::string> _partition_column_names;
 };