… be/src/exec/CMakeLists.txt
… be/src/exec/iceberg/iceberg_delete_builder.cpp
… be/src/exec/iceberg/iceberg_delete_builder.h
… be/src/exec/mor_processor.cpp
… be/src/exec/mor_processor.h
… fe/fe-core/src/main/java/com/starrocks/catalog/IcebergTable.java
… fe/fe-core/src/main/java/com/starrocks/planner/IcebergEqualityDeleteScanNode.java
… fe/fe-core/src/main/java/com/starrocks/planner/IcebergScanNode.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/LogicalPlanPrinter.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/OptExpressionVisitor.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Optimizer.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/OperatorType.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/OperatorVisitor.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalIcebergEqualityDeleteScanOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalIcebergEqualityDeleteScanOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleSet.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleType.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/IcebergEqualityDeleteScanImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/IcebergEqualityDeleteRewriteRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/OnPredicateMoveAroundRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateScanRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/statistics/StatisticsCalculator.java
… fe/fe-core/src/main/java/com/starrocks/sql/plan/PlanFragmentBuilder.java
… fe/fe-core/src/test/java/com/starrocks/common/util/ProfilingExecPlanTest.java
… fe/fe-core/src/test/java/com/starrocks/planner/IcebergScanNodeTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/IcebergEqualityDeletePlanTest.java
… gensrc/thrift/PlanNodes.thrift