## Why I'm doing:
The current implementation of Iceberg reading eq-delete file for mor is to do a local left anti join in each scanner thread in units of scan range.
There are three problems with this:
- A large data file such as 1G may have multiple splits, each of which repeatedly reads the same delete file.
- A delete file may match different data files, causing the same delete file to be read multiple times.
- Even if data cache is used for delete files, the parsed delete file is used as the right table to build a hash table. In extreme cases, memory usage will be very large. If there are many delete files for users, the query often OOM. Many cases can only run normally under one concurrency.

## What I'm doing:

This patch implements iceberg equality deletes as a join, rather than reading the data file as the left table and the delete file as the right table for local left anti join. This optimization replaces the previous solution of using local hash joiner in each scanner thread. Compared to the previous solution, the main purpose is to reduce the overhead of repeatedly reading delete files and repeatedly building hashtable since a iceberg equality delete file may be matched by many data files after iceberg planning. This rule needs to strictly meet the check requirements before it can be rewritten.

Three are three conditions that need to rewrite:
- iceberg format is v2 format
- snapshot summary exist equality delete file.
- exists real delete file in scan task after iceberg job planning. 

We'll rewrite three patterns.
The first common case:
iceberg identifier column (also the same as pk) are `identifier_col` and `p1`.
```
mysql> explain select * from pk_int_50_par_int_50;
+---------------------------------------------------------------------------------+
| Explain String                                                                  |
+---------------------------------------------------------------------------------+
| PLAN FRAGMENT 0                                                                 |
|  OUTPUT EXPRS:1: identifier_col | 2: data | 3: par_col                          |
|   PARTITION: UNPARTITIONED                                                      |
|                                                                                 |
|   RESULT SINK                                                                   |
|                                                                                 |
|   5:EXCHANGE                                                                    |
|                                                                                 |
| PLAN FRAGMENT 1                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: RANDOM                                                             |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 05                                                             |
|     UNPARTITIONED                                                               |
|                                                                                 |
|   4:Project                                                                     |
|   |  <slot 1> : 1: identifier_col                                               |
|   |  <slot 2> : 2: data                                                         |
|   |  <slot 3> : 3: par_col                                                      |
|   |                                                                             |
|   3:HASH JOIN                                                                   |
|   |  join op: LEFT ANTI JOIN (BROADCAST)                                        |
|   |  colocate: false, reason:                                                   |
|   |  equal join conjunct: 1: identifier_col = 5: identifier_col                 |
|   |  equal join conjunct: 3: par_col = 6: par_col                               |
|   |  other join predicates: 4: $data_sequence_number < 7: $data_sequence_number |
|   |                                                                             |
|   |----2:EXCHANGE                                                               |
|   |                                                                             |
|   0:IcebergScanNode                                                             |
|      TABLE: pk_int_50_par_int_50                                                |
|      cardinality=119750000                                                      |
|      avgRowSize=4.0                                                             |
|                                                                                 |
| PLAN FRAGMENT 2                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: RANDOM                                                             |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 02                                                             |
|     UNPARTITIONED                                                               |
|                                                                                 |
|   1:IcebergEqualityScanNode                                                     |
|      TABLE: pk_int_50_par_int_50_eq_delete_identifier_col_par_col               |
|      cardinality=497500                                                         |
|      avgRowSize=3.0                                                             |
|      Iceberg identifier columns: [identifier_col, par_col]                      |
+---------------------------------------------------------------------------------+
```

The second case with mutable pk column:
pk column before altering the table is `k1`.
pk column after altering the table is `k1`.
```
mysql> explain select * from test_k1_and_k2_ab;
+---------------------------------------------------------------------------------+
| Explain String                                                                  |
+---------------------------------------------------------------------------------+
| PLAN FRAGMENT 0                                                                 |
|  OUTPUT EXPRS:1: k1 | 2: k2                                                     |
|   PARTITION: UNPARTITIONED                                                      |
|                                                                                 |
|   RESULT SINK                                                                   |
|                                                                                 |
|   11:EXCHANGE                                                                   |
|                                                                                 |
| PLAN FRAGMENT 1                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: HASH_PARTITIONED: 4: k2                                            |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 11                                                             |
|     UNPARTITIONED                                                               |
|                                                                                 |
|   10:Project                                                                    |
|   |  <slot 1> : 1: k1                                                           |
|   |  <slot 2> : 2: k2                                                           |
|   |                                                                             |
|   9:HASH JOIN                                                                   |
|   |  join op: RIGHT ANTI JOIN (PARTITIONED)                                     |
|   |  colocate: false, reason:                                                   |
|   |  equal join conjunct: 4: k2 = 2: k2                                         |
|   |  other join predicates: 3: $data_sequence_number < 5: $data_sequence_number |
|   |                                                                             |
|   |----8:EXCHANGE                                                               |
|   |                                                                             |
|   1:EXCHANGE                                                                    |
|                                                                                 |
| PLAN FRAGMENT 2                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: HASH_PARTITIONED: 1: k1                                            |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 08                                                             |
|     HASH_PARTITIONED: 2: k2                                                     |
|                                                                                 |
|   7:Project                                                                     |
|   |  <slot 1> : 1: k1                                                           |
|   |  <slot 2> : 2: k2                                                           |
|   |  <slot 3> : 3: $data_sequence_number                                        |
|   |                                                                             |
|   6:HASH JOIN                                                                   |
|   |  join op: LEFT ANTI JOIN (PARTITIONED)                                      |
|   |  colocate: false, reason:                                                   |
|   |  equal join conjunct: 1: k1 = 6: k1                                         |
|   |  other join predicates: 3: $data_sequence_number < 7: $data_sequence_number |
|   |                                                                             |
|   |----5:EXCHANGE                                                               |
|   |                                                                             |
|   3:EXCHANGE                                                                    |
|                                                                                 |
| PLAN FRAGMENT 3                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: RANDOM                                                             |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 05                                                             |
|     HASH_PARTITIONED: 6: k1                                                     |
|                                                                                 |
|   4:IcebergEqualityScanNode                                                     |
|      TABLE: test_k1_and_k2_ab_eq_delete_k1                                      |
|      cardinality=3                                                              |
|      avgRowSize=2.0                                                             |
|      Iceberg identifier columns: [k1]                                           |
|                                                                                 |
|                                                                                 |
| PLAN FRAGMENT 4                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: RANDOM                                                             |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 03                                                             |
|     HASH_PARTITIONED: 1: k1                                                     |
|                                                                                 |
|   2:IcebergScanNode                                                             |
|      TABLE: test_k1_and_k2_ab                                                   |
|      cardinality=5                                                              |
|      avgRowSize=3.0                                                             |
|                                                                                 |
| PLAN FRAGMENT 5                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: RANDOM                                                             |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 01                                                             |
|     HASH_PARTITIONED: 4: k2                                                     |
|                                                                                 |
|   0:IcebergEqualityScanNode                                                     |
|      TABLE: test_k1_and_k2_ab_eq_delete_k2                                      |
|      cardinality=3                                                              |
|      avgRowSize=2.0                                                             |
|      Iceberg identifier columns: [k2]                                           |
+---------------------------------------------------------------------------------+
94 rows in set (0.16 sec)
```

The third case with partition evolution.

Partition Table With 1 delete schema: [k1, p1], Partition column: [p1]. Write some records to this table.
Then alter table partition field (partition evolution): (p1 -> bucket(5, p1)). Write some records to this table.
```
mysql> explain select * from test_bucket_table;
+---------------------------------------------------------------------------------+
| Explain String                                                                  |
+---------------------------------------------------------------------------------+
| PLAN FRAGMENT 0                                                                 |
|  OUTPUT EXPRS:1: k1 | 2: k2 | 3: p1                                             |
|   PARTITION: UNPARTITIONED                                                      |
|                                                                                 |
|   RESULT SINK                                                                   |
|                                                                                 |
|   6:EXCHANGE                                                                    |
|                                                                                 |
| PLAN FRAGMENT 1                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: HASH_PARTITIONED: 1: k1, 3: p1, 5: $spec_id                        |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 06                                                             |
|     UNPARTITIONED                                                               |
|                                                                                 |
|   5:Project                                                                     |
|   |  <slot 1> : 1: k1                                                           |
|   |  <slot 2> : 2: k2                                                           |
|   |  <slot 3> : 3: p1                                                           |
|   |                                                                             |
|   4:HASH JOIN                                                                   |
|   |  join op: LEFT ANTI JOIN (PARTITIONED)                                      |
|   |  colocate: false, reason:                                                   |
|   |  equal join conjunct: 1: k1 = 6: k1                                         |
|   |  equal join conjunct: 3: p1 = 7: p1                                         |
|   |  equal join conjunct: 5: $spec_id = 9: $spec_id                             |
|   |  other join predicates: 4: $data_sequence_number < 8: $data_sequence_number |
|   |                                                                             |
|   |----3:EXCHANGE                                                               |
|   |                                                                             |
|   1:EXCHANGE                                                                    |
|                                                                                 |
| PLAN FRAGMENT 2                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: RANDOM                                                             |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 03                                                             |
|     HASH_PARTITIONED: 6: k1, 7: p1, 9: $spec_id                                 |
|                                                                                 |
|   2:IcebergEqualityScanNode                                                     |
|      TABLE: test_bucket_table_eq_delete_k1_p1                                   |
|      cardinality=2                                                              |
|      avgRowSize=4.0                                                             |
|      Iceberg identifier columns: [k1, p1]                                       |
|                                                                                 |
|                                                                                 |
| PLAN FRAGMENT 3                                                                 |
|  OUTPUT EXPRS:                                                                  |
|   PARTITION: RANDOM                                                             |
|                                                                                 |
|   STREAM DATA SINK                                                              |
|     EXCHANGE ID: 01                                                             |
|     HASH_PARTITIONED: 1: k1, 3: p1, 5: $spec_id                                 |
…