@stephen-shelby 2025-02-08
> @Seaven not sure if it's the right way to ask previously we had a mechanism for iceberg scan inspection using
> 
> ```
> set enable_profile='true';
> trace values select count(*) from table;
> ```
> 
> now this mechanism no longer exists
> 
> we have something similar in the log output
> 
> ```
> 2025-02-07 12:59:37.423Z INFO (com.starrocks.connector.hive.ConnectorTableMetadataProcessor|37) [LoggingMetricsReporter.report():38] Received metrics report: ScanReport{tableName=xxx, snapshotId=5729558768678613989, filter=true, schemaId=0, projectedFieldIds=[2, 3, 5, 6, 7, 8, 9, 10, 11], projectedFieldNames=[record_count, file_count, position_delete_record_count, position_delete_file_count, equality_delete_record_count, equality_delete_file_count, last_updated_at, last_updated_snapshot_id, total_data_file_size_in_bytes], scanMetrics=ScanMetricsResult{totalPlanningDuration=TimerResult{timeUnit=NANOSECONDS, totalDuration=PT0.01853216S, count=1}, resultDataFiles=CounterResult{unit=COUNT, value=0}, resultDeleteFiles=CounterResult{unit=COUNT, value=0}, totalDataManifests=CounterResult{unit=COUNT, value=0}, totalDeleteManifests=CounterResult{unit=COUNT, value=0}, scannedDataManifests=CounterResult{unit=COUNT, value=0}, skippedDataManifests=CounterResult{unit=COUNT, value=0}, totalFileSizeInBytes=CounterResult{unit=BYTES, value=0}, totalDeleteFileSizeInBytes=CounterResult{unit=BYTES, value=0}, skippedDataFiles=CounterResult{unit=COUNT, value=0}, skippedDeleteFiles=CounterResult{unit=COUNT, value=0}, scannedDeleteManifests=CounterResult{unit=COUNT, value=0}, skippedDeleteManifests=CounterResult{unit=COUNT, value=0}, indexedDeleteFiles=CounterResult{unit=COUNT, value=0}, equalityDeleteFiles=CounterResult{unit=COUNT, value=0}, positionalDeleteFiles=CounterResult{unit=COUNT, value=0}}, metadata={iceberg-version=Apache Iceberg 1.7.1 (commit 4a432839233f2343a9eae8255532f911f06358ef)}}
> ```
> 
> but all the counts are zero even though the session had set enable_profile='true' before select
> 
> can you please bring back the original trace values mechanism back?

@Youngwb @dirtysalt please take a look if you have time