@dirtysalt 2025-02-08
@eshishki 

in 3.4.0 we introduced a new mechansim to dispatch files. Instead of collecting files at one time, we collecting files in multiple rounds. And by doing that, we are not enable to collect those stats.

So to disable that, you can set
> set enable_connector_incremental_scan_ranges = false;

And then you can see the trace values.