@eshishki 2025-02-08
@dirtysalt thanks, with this option i can see ScanMetricsResult in trace values output
i presume this should only be off for debugging, right?