@dirtysalt 2025-02-09
> @dirtysalt thanks, with this option i can see ScanMetricsResult in trace values output i presume this should only be off for debugging, right?

Yes.  But internally we don't actually have debug mode or not.  

And trace values is just for internal use. No guarantee of backward compitiability.